backend/fastapi-document-retrieval/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
milvus
python-multipart
alembic
pytest
pytest-asyncio
asyncpg
databases